+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # 0.3.0 (Unreleased)
 
+ENHANCEMENTS:
+
+* resource/ec_deployment_traffic_filter: Add support for Azure Private Link traffic rules. ([#340](https://github.com/elastic/terraform-provider-ec/issues/340))
+
 # 0.2.1 (Jun 17, 2021)
 
 BUG FIXES: